add check for new marker to KASP and Intertek

+++ lib/SGN/Controller/AJAX/GenotypesVCFUpload.pm
@@ -413,22 +413,27 @@ sub upload_genotype_verify_POST : Args(0) {
         }
     }
 
-    my $organism_q = "SELECT organism_id FROM organism WHERE species = ?";
-    my @found_organisms;
-    my $h = $schema->storage->dbh()->prepare($organism_q);
-    $h->execute($organism_species);
-    while (my ($organism_id) = $h->fetchrow_array()){
-        push @found_organisms, $organism_id;
+    my $organism_id;
+    if ($organism_species) {
+        my $organism_q = "SELECT organism_id FROM organism WHERE species = ?";
+        my @found_organisms;
+        my $h = $schema->storage->dbh()->prepare($organism_q);
+        $h->execute($organism_species);
+        while (my ($organism_id) = $h->fetchrow_array()){
+            push @found_organisms, $organism_id;
+        }
+        if (scalar(@found_organisms) == 0){
+            $c->stash->{rest} = { error => 'The organism species you provided is not in the database! Please contact us.' };
+            $c->detach();
+        }
+        if (scalar(@found_organisms) > 1){
+            $c->stash->{rest} = { error => 'The organism species you provided is not unique in the database! Please contact us.' };
+            $c->detach();
+        }
+        $organism_id = $found_organisms[0];
+    } else {
+	print STDERR "organism species not defined\n";
     }
-    if (scalar(@found_organisms) == 0){
-        $c->stash->{rest} = { error => 'The organism species you provided is not in the database! Please contact us.' };
-        $c->detach();
-    }
-    if (scalar(@found_organisms) > 1){
-        $c->stash->{rest} = { error => 'The organism species you provided is not unique in the database! Please contact us.' };
-        $c->detach();
-    }
-    my $organism_id = $found_organisms[0];
 
     my $parser;
     if ($upload_original_name) {
@@ -696,27 +701,42 @@ sub upload_genotype_verify_POST : Args(0) {
                 nd_protocol_id => $protocol_id
             });
             my $stored_markers = $stored_protocol->markers();
-
-            while (my ($marker_name, $marker_obj) = each %$stored_markers) {
-                while (my ($chrom, $new_marker_data_1) = each %$new_marker_data) {
-                    if ($new_marker_data_1->{$marker_name}) {
-                        my $protocol_data_obj = $new_marker_data_1->{$marker_name};
-                        while (my ($key, $value) = each %$marker_obj) {
-                            if ($value ne $protocol_data_obj->{$key}) {
-                                push @protocol_match_errors, "Marker $marker_name in the previously loaded protocol has $value for $key, but in your file now shows ".$protocol_data_obj->{$key};
+	    my @all_stored_markers = keys %$stored_markers;
+	    my %compare_marker_names = map {$_ => 1} @all_stored_markers;
+	    my $total_marker_count = 0;
+            my @mismatch_marker_names;
+            my @mismatch_markers;
+            while (my ($chrom, $new_marker_data_1) = each %$new_marker_data) {
+                while (my ($marker_name, $new_marker_details) = each %$new_marker_data_1) {
+                    $total_marker_count++;
+                    if (exists($compare_marker_names{$marker_name})) {
+                        while (my ($key, $value) = each %$new_marker_details) {
+                            if ($value ne ($stored_markers->{$marker_name}->{$key})) {
+                                push @protocol_match_errors, "Marker $marker_name in your file has $value for $key, but in the previously stored protocol shows ".$stored_markers->{$marker_name}->{$key};
                             }
                         }
+                    } else {
+                        push @mismatch_marker_names, $marker_name;
+                        push @mismatch_markers, [$chrom, $marker_name];
                     }
                 }
             }
 
-            if (scalar(@protocol_match_errors) > 0){
-                my $protocol_warning;
-                foreach my $match_error (@protocol_match_errors) {
-                    $protocol_warning .= $match_error."<br>";
-                }
-                if (!$accept_warnings){
-                    $c->stash->{rest} = { warning => $protocol_warning };
+            if (scalar(@protocol_match_errors)){
+		if ($add_markers) {
+		    if ($total_marker_count && (scalar(@mismatch_marker_names) / $total_marker_count) < 0.1) {
+                        print STDERR "Adding new markers\n";
+                        $store_genotypes->store_new_markers_in_protocolprop(\@mismatch_markers);
+                    } else {
+		        $c->stash->{rest} = { error => "Too many new markers"};
+                        $c->detach();
+                    }
+		} else {
+		    my $marker_name_error = "<br>";
+                    foreach my $error ( sort @mismatch_marker_names) {
+                        $marker_name_error .= $error."<br>";
+                    }
+                    $c->stash->{rest} = { error => "These marker names in your file are not in the selected protocol. $marker_name_error"};
                     $c->detach();
                 }
             }
@@ -798,7 +818,7 @@ sub upload_genotype_verify_POST : Args(0) {
 
     } else {
         print STDERR "Parser plugin $parser_plugin not recognized!\n";
-        $c->stash->{rest} = { error => "Parser plugin $parser_plugin not recognized!" };
+	#$c->stash->{rest} = { error => "Parser plugin $parser_plugin not recognized!" };
         $c->detach();
     }
